Curated Submission 2 (p5.js)
For my second curated submission is from week 4 using p5.js. 
This is when we began to explore making loops and patterns. How you can make something that may seem hard and make it more similar by having one thing repeat itself rather than having multiple of the same item. I choose this code because it has the randomness of colour which I like and also the random lines being layered through adding texture to the colour. Which i think looks very pretty you also get some randomisation which we did in week 3. Both the colour and lines when you pause the code will be different every time which is also interesting.