Curated Submission 2 (p5.js)
Curated Submission 1 (p5.js)
My first curated submission is from week 2 on p5.js. We were learning how to write simple javascript and how the mouse can interact with our code.
The reason why I choose this code is I was finally beginning to understand that code can be more than just for technical reasons but it can also be used to make art and to express yourself. I also was beging to understand how code can work together.
I chose to make a dog because I was missing my dog from back home and I also thought it was simplistic shape that I should be able to replicate on my code pretty easily.
How the code works is by using random() in a loop to draw a vertical line made of different colours, with each vertex(x, y) placed randomly across the width of the canvas.