Curated Submission 5 (Node-based coding)
(In project 2)
For my fifth curated submission, I chose a project from Week 7 where we did node-based programming. The focus was on using TOPs (Texture Operators) in TouchDesigner. This week I was just playing around and beginning to understand how the coding works for the new software. I ended up working with 2D image and played around with the visuals using things like patterns, noise, and compositing. I found this approach really engaging because it allowed me to experiment with layering textures, adjusting colour, and applying effects like blur and displacement. This was pretty much me jsut seeing what certain things did. The only issue is I did too much to the code that I don't know how the code works in the sense I don't know how to repeat what I have made, which is both a good and a bad thing. I know how each node works independently, but how the work together to get the end product is something that I still struggle to get my head around particularly with this code, which is apart of why I choose it. I think the code is aesthetically pleasing to look at and its really interesting to see how the code developed.