Curated Submission 6 (Node-based programming)
My sixth curated submission is from Week 8 where we used node-based programming. This week we explored feedback loops in TouchDesigner. This code focuses on creating a "pattern" that builds and changes over time, using a feedback network to loop previous frames back into the current frame. I found this really fascinating because it creates a sense of motion and transformation that feels fluid and its like a "living" code as it keeps evolving. In my code, I combined basic shapes, squares, with transformation and blur TOPs inside a feedback loop, allowing the visuals to shift and distort gradually. This allows the pattern to react to itself, producing an almost infinite square pattern repeating itself. I followed a video guide for this particular code it was interesting to learn outside of the lecture setting as it was harder to understand how or why the code works. That is why I choose this code, because I followed the video and it didn't work quite turn out right. I know I am missing a few things but it still works to some extend and it was fun experimenting and learning in a environment where I may not be as comfortable.